Temporary fix while chromium fixes badge text color behavior
A badge's text color is decided automatically by the browser, the
extension cannot control this behavior. If the badge's background color
is dark (as it normally is for this extension), the browser is supposed
to pick a light color for the badge's text. This doesn't happen anymore,
so for the time being, a lighter background color has been picked so
that the badge text is readable again.